Clean query from user
