+++ discovery_engine_core/ai/ai/Cargo.toml
@@ -19,7 +19,6 @@ ndarray = "=0.15.3"
 obake = { git = "https://github.com/doctorn/obake", rev = "d6bea07e355ca4adf353c1e627f13c8c3286361b" }
 rand = "0.8.5"
 rayon = "1.5.1"
-regex = { version = "1.5.5", features = ["unicode-gencat"] }
 serde = { version = "1.0.136", features = ["derive", "rc"] }
 serde_repr = "0.1.7"
 smallvec = "1.8.0"
