+++ discovery_engine_core/providers/Cargo.toml
@@ -11,6 +11,7 @@ displaydoc = "0.2.3"
 itertools = "0.10.3"
 lazy_static = "1.4.0"
 maplit = "1.0.2"
+regex = { version = "1.5.5", features = ["unicode-gencat"] }
 reqwest = { version = "0.11.9", features = ["json", "rustls-tls"], default-features = false }
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.79"
